Installed the UmbracoCms.Core NuGet package
Since Umbraco wasn't installed as a NuGet package, the assembly references would point to non-existing DLLs after doing a clone from the repository. I have (re)installed the same versions as the references were pointing to, so this shouldn't cause any issues.

However by installing the UmbracoCms.Core NuGet package, this will show up as a depedency when building the NuGet package for this project. But luckily there is a way to handle that if you wish to avoid the dependency ;)